More refactoring

+++ Sadie.Game.Players/IPlayerRepository.cs
@@ -4,15 +4,15 @@ namespace Sadie.Game.Players;
 
 public interface IPlayerRepository : IAsyncDisposable
 {
-    bool TryGetPlayerById(long id, out IPlayer? player);
+    bool TryGetPlayerById(int id, out IPlayer? player);
     bool TryGetPlayerByUsername(string username, out IPlayer? player);
     Task<Tuple<bool, IPlayer?>> TryGetPlayerBySsoAsync(INetworkObject networkObject, string sso);
     bool TryAddPlayer(IPlayer player);
-    Task<bool> TryRemovePlayerAsync(long playerId);
-    Task MarkPlayerAsOnlineAsync(long id);
+    Task<bool> TryRemovePlayerAsync(int playerId);
+    Task MarkPlayerAsOnlineAsync(int id);
     Task MarkPlayerAsOfflineAsync(IPlayer player);
-    Task ResetSsoTokenForPlayerAsync(long id);
+    Task ResetSsoTokenForPlayerAsync(int id);
     int Count();
-    Task<Tuple<bool, IPlayerData?>> TryGetPlayerData(long playerId);
+    Task<Tuple<bool, IPlayerData?>> TryGetPlayerData(int playerId);
     Task<Tuple<bool, IPlayerData?>> TryGetPlayerDataByUsername(string username);
 }
