+++ Sadie.Game.Players/PlayerRepository.cs
@@ -8,16 +8,16 @@ public class PlayerRepository : IPlayerRepository
 {
     private readonly ILogger<PlayerRepository> _logger;
     private readonly IPlayerDao _playerDao;
-    private readonly ConcurrentDictionary<long, IPlayer> _players;
+    private readonly ConcurrentDictionary<int, IPlayer> _players;
 
     public PlayerRepository(ILogger<PlayerRepository> logger, IPlayerDao playerDao)
     {
         _logger = logger;
         _playerDao = playerDao;
-        _players = new ConcurrentDictionary<long, IPlayer>();
+        _players = new ConcurrentDictionary<int, IPlayer>();
     }
 
-    public bool TryGetPlayerById(long id, out IPlayer? player)
+    public bool TryGetPlayerById(int id, out IPlayer? player)
     {
         return _players.TryGetValue(id, out player);
     }
@@ -35,7 +35,7 @@ public class PlayerRepository : IPlayerRepository
 
     public bool TryAddPlayer(IPlayer player) => _players.TryAdd(player.Id, player);
 
-    public async Task<bool> TryRemovePlayerAsync(long playerId)
+    public async Task<bool> TryRemovePlayerAsync(int playerId)
     {
         var result = _players.TryRemove(playerId, out var player);
 
@@ -50,7 +50,7 @@ public class PlayerRepository : IPlayerRepository
         return result;
     }
 
-    public async Task MarkPlayerAsOnlineAsync(long id)
+    public async Task MarkPlayerAsOnlineAsync(int id)
     {
         await _playerDao.MarkPlayerAsOnlineAsync(id);
     }
@@ -60,7 +60,7 @@ public class PlayerRepository : IPlayerRepository
         await _playerDao.MarkPlayerAsOfflineAsync(player);
     }
 
-    public async Task ResetSsoTokenForPlayerAsync(long id)
+    public async Task ResetSsoTokenForPlayerAsync(int id)
     {
         await _playerDao.ResetSsoTokenForPlayerAsync(id);
     }
@@ -70,7 +70,7 @@ public class PlayerRepository : IPlayerRepository
         return _players.Count;
     }
 
-    public async Task<Tuple<bool, IPlayerData?>> TryGetPlayerData(long playerId)
+    public async Task<Tuple<bool, IPlayerData?>> TryGetPlayerData(int playerId)
     {
         return await _playerDao.TryGetPlayerData(playerId);
     }
